Update upload-model-to-dockerhub.yml (#39)
Added two new steps to the workflow, namely, buildForTest, and testBuiltImage.

The `buildForTest` step only builds an amd64 image with the latest tag since we test this image through the ersilia CLI which only looks for the latest tag.

The `testBuiltImage` serves the newly built model image and tests it against a simple input. Then I use jq to count the fields within the model outcome followed by checking for null values in the output. If there are any null values within the output this test fails and the model push will not succeed.

Upon success of this test, the subsequent build stage will only build the arm64 image and reuse the previously built amd64 image and push them both to DockerHub.

Also moved the installation of ersilia a few steps up to use within the testBuiltImage stage